refactor: extract common scroller logic to reusable mixin
